+++ header/Tetramino.hpp
@@ -11,12 +11,12 @@ class Tetramino: public sf::Drawable
 private:
 	static const unsigned int numOfBlocks = 4;
 	Block blocks[numOfBlocks];
-	sf::Vector2f offset;
-	sf::Vector2f buildingPosition;
+	sf::Vector2i offset;
+	sf::Vector2i buildingPosition;
 	float speed; //number of steps down per second
 	float elapsedTime;
 
-	float moveValueX;
+	int moveValueX;
 	bool isRotate;
 	bool isAccelerate;
 
@@ -27,13 +27,13 @@ public:
 	Block& operator[](const unsigned int index);
 	~Tetramino();
 
-	void setPositionOffset(const sf::Vector2f position);
-	void setBuildingPosition(const sf::Vector2f position);
-	sf::Vector2f getBuildingPosition() const;
-	void setModel(	const sf::Vector2f pos0, 
-					const sf::Vector2f pos1, 
-					const sf::Vector2f pos2, 
-					const sf::Vector2f pos3);
+	void setPositionOffset(const sf::Vector2i position);
+	void setBuildingPosition(const sf::Vector2i position);
+	sf::Vector2i getBuildingPosition() const;
+	void setModel(	const sf::Vector2i pos0, 
+					const sf::Vector2i pos1, 
+					const sf::Vector2i pos2, 
+					const sf::Vector2i pos3);
 	void adaptPixelPosition();
 	void setColor(const sf::Color color);
 	void setSpeed(const float value);
@@ -46,7 +46,7 @@ public:
 	void speedDown();
 	void update(const float deltaTime);
 
-	void hardMove(const sf::Vector2f step);
+	void hardMove(const sf::Vector2i step);
 
 	void draw(sf::RenderTarget &target, sf::RenderStates states = sf::RenderStates::Default) const;
 private:
